+++ Changelog.md
@@ -5,6 +5,7 @@
 
 - Added zones with scopes (shipping, billing, etc) and members (province, country)
 - Added `country`, `province` and `zone` models to the Relation morphMap
+- Added the `Zones` query class (eg. `Zones::withShippingScope()->theCountryBelongsTo('CA')`)
 
 ## 2.4.0
 ##### 2023-02-16